feat(arch): SymbolQueryPipeline (3 variants) + #318 transport envelope (bulk)
Bulk integration of the SymbolQuery pipeline series (PR-A through PR-E,
originally PRs #317 / #319 / #320 / #321 / #322) plus the #318
dogfood fix for the transport-layer session envelope. Closing the
five-PR stack as a single squash because the stacked-PR rebase chain
collided with main updates from the parallel #314/#315/#316 merges,
and the final tree is identical.

Net effect:
  - `crates/codelens-mcp/src/tools/symbol_query/` — new deep module.
    Three variants of `SymbolQueryRequest`: `RankedContext`,
    `FindSymbol`, `SymbolsOverview`. Each handler in
    `tools/symbols/handlers.rs` is now a 3-line entry that
    constructs a request and calls `SymbolQueryPipeline::run`.
    1170+ LOC of orchestration + stage helpers + tests concentrate
    here.
  - `crates/codelens-mcp/src/tools/semantic_retriever.rs` — new
    cross-cutting seam for `semantic_status` /
    `semantic_results_for_query`. Two adapters (symbol-query
    pipeline + impact-report family) → real seam.
  - `crates/codelens-mcp/src/tools/symbols/handlers.rs` — shrinks
    from 1575 LOC (god module) to 570 LOC (-64%). Only
    bm25_symbol_search, sparse_symbol_hits_for_query,
    adapt_budget_to_context_window, refresh_symbol_index,
    get_complexity, get_project_structure, search_symbols_fuzzy,
    flatten_symbols, suggested_follow_up, confidence_tier, and the
    three pipeline thin entries remain.
  - `crates/codelens-mcp/src/tools/symbols/fusion.rs` — deleted
    (rank_fusion_policy moved into
    `symbol_query/ranked_context.rs`).
  - `crates/codelens-mcp/src/tools/symbols/analyzer.rs` — shrinks
    to just `semantic_scores_for_query` (5 rank-fusion helpers
    moved into the pipeline module).
  - CLAUDE.md — new "Symbol-query path lives behind one seam"
    section under "Three concepts that show up across files".
    Codifies the visibility policy: `pub(crate)` helpers spanning
    pipeline + non-pipeline handler stay where they are; stage
    helpers stay file-private inside the pipeline module.

#318 fix (transport-layer envelope completeness):
  - `transport_http.rs::unknown_session_response()`: new helper
    returning the structured envelope + `x-codelens-session-rotate: 1`
    header.
  - Both session-miss paths (POST /mcp non-initialize validation +
    GET /mcp SSE session lookup) now funnel through this helper.
    Previously only the legacy fix-up covered one path while the
    SSE path emitted raw "Unknown session" text, so reinitialize
    hints never reached well-behaved clients on hot-reload of the
    SCIP index.

Build matrix:
  - cargo check --workspace --features http,semantic,scip-backend ✓
  - cargo check --workspace --no-default-features ✓
  - cargo clippy --features semantic -- -D warnings ✓
  - cargo fmt --all -- --check ✓
  - cargo test -p codelens-mcp --bin codelens-mcp
    --features http,semantic,scip-backend
    → 714 passed / 7 ignored / 0 failed (51.62s)

Closes #318. Supersedes #319, #320, #321, #322, and #309
(unknown_session_response envelope from #309 is included here so
the SSE path picks it up too).
