+++ crates/codelens-mcp/src/tools/mod.rs
@@ -19,6 +19,7 @@ pub(crate) mod semantic_edit_args;
 pub(crate) mod semantic_retriever;
 pub mod session;
 mod suggestions;
+pub(crate) mod symbol_query;
 pub mod symbols;
 pub mod workflows;
